requirements.txt
# PyTorch with CUDA 12.1 support (optimized for RTX 4090 and RTX 5090)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Additional dependencies
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.7.0
requests>=2.28.0
xformers>=0.0.20
flask>=2.3.0
flask-cors>=3.0.10
websocket-client>=1.6.0
pyotp>=2.9.0
qrcode>=7.4.2
authlib>=1.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
